+++ horaedb-protocol/src/main/java/org/apache/horaedb/BlockingStreamIterator.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2023 CeresDB Project Authors. Licensed under Apache-2.0.
  */
-package org.apache;
+package org.apache.horaedb;
 
 import java.util.Iterator;
 import java.util.concurrent.BlockingQueue;
