+++ horaedb-protocol/src/main/java/org/apache/horaedb/util/Utils.java
@@ -25,7 +25,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 
-import org.apache.Route;
+import org.apache.horaedb.Route;
 
 import org.apache.horaedb.models.Err;
 import org.apache.horaedb.models.Keyword;
