+++ horaedb-protocol/src/test/java/org/apache/horaedb/HoraeDBClientTest.java
@@ -7,9 +7,6 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
-import org.apache.HoraeDBClient;
-import org.apache.RouteMode;
-import org.apache.WriteClient;
 import org.apache.horaedb.common.util.MetricsUtil;
 import org.apache.horaedb.common.util.internal.ReferenceFieldUpdater;
 import org.apache.horaedb.common.util.internal.Updaters;
